more body tweaks
had the bright idea to actually measure some seats and make sure there's room for the occupants, which brought the roof line up. while I was at it I elongated the body to give it some extra leg room and engine bay capacity, bringing it to the overall bounding dimensions of a normal subcompact. tweaked the roof line again into a kammback, though the fin lights stay to maintain that personality. yet more fender tweaks, since it was looking like an old-school torpedo from the top down, and not in a flattering way. probably going to start a new body from scratch, just to see where the inspiration leads.