housekeeping and further tweaks
went ahead and sorted out my files into their own subfolder, along with splitting the renders from the models just to keep things a little tidier. also added glass to the car and as an experiment made the windscreen a bit more horizontal, though the jury's still out on that. speaking of, I've got jury duty on Monday so I'm going to try and bang out a bit more with this before taking a brief stay of absence this coming week.