more fender work, headlight experiments
forgot to commit yesterday and didn't get a lot more done today. working on closed fenders - not super happy with how they're turning out, probably going to take a fresh run at it and see what happens with a blank slate. 01211 also has the headlights on the fenders which doesn't quite look good, but the idea of headlights turning with steering is a good idea that's potentially worth thinking about again in the future.