more fender work
Tried something completely different, not really the classic styling but not bad. Fully enclosed floating fenders are really hard to make look halfway decent, and there's not many examples to take inspiration from. While the lower CoD is definitely something we still want to consider, there are also merits to open fenders - easier access to the lugs for everyman consumers, and (negligible) downforce via the magnus effect for the guys that care about that.