sorted out some stuff
the v0.1 model folder was getting a bit outrageous with the revisions to the fenders, so i added a subfolder for the v0.1.2 models specifically. I also grabbed a workspace to try starting a v0.3 body, but right now that's just wheels and lightroom to position things when i get started.