light organization
crediting each contributor's works and setting them in the Initial Designs folder, for easier navigation in the future. Debated between standardizing the organization and leaving it to individual contributors' workflows, ultimately decided on the latter, at least for now. This content is going to be offloaded to the body-design branch for now, to keep the master branch a little more clean.